+++ pkgs/xsoulspace_review/lib/xsoulspace_review_foundation.dart
@@ -1,8 +1,10 @@
 export 'package:wiredash/wiredash.dart';
+export 'package:xsoulspace_logger/xsoulspace_logger.dart';
 export 'package:xsoulspace_review_interface/xsoulspace_review_interface.dart';
 
 export 'src/consent_screen.dart';
+export 'src/logger_config.dart';
+export 'src/logger_extensions.dart';
 export 'src/review_foundation.dart';
 export 'src/store_review_requester.dart';
 export 'src/user_feedback/user_feedback.dart';
-